shems-backend/requirement.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
python-dotenv
pandas
scikit-learn
joblib
pandas
scikit-learn
joblib